feat: add context provider and PropertiesControlPanel component
The provider is used to manage the state of the application. It contains the state of the current component, the list of components, the image, the color, the rotation, the position, the width, the height, the opacity, the z-index, the padding, the font size, the font weight, the text, the radius, and the show status.

The provider also contains methods to set the state, add an image, add a text, move an element, resize an element, rotate an element, remove a component, and remove the background.

The create component is used to create a new component. It takes the name and the type of the component as parameters.