feat: convert http to https for image_url
This commit converts all http image_urls to https image_urls.

This is done for security reasons, as https is more secure than http. By converting all image_urls to https, we are ensuring that all images are loaded securely.

This change also updates the cloudinary uploader to use secure uploads. This ensures that all images uploaded to cloudinary are uploaded securely.

(main)